requirement.txt
anthropic==0.39.0
datasets==3.2.0
google-generativeai==0.8.3
json-repair==0.30.1
httpx==0.27.2
ipywidgets==8.1.5
numpy==2.0.2
pandarallel==1.6.5
pandas==2.2.3
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.88.0
tenacity==9.0.0
tiktoken==0.8.0
litellm==1.72.6
fire==0.7.0
streamlit==1.45.1